fix(account_id): ensure account_id is populated in vars
Why:

* Currently only gets populated with stack refs. Force early
  check.

+++ stk/cfn_bucket.py
@@ -36,6 +36,8 @@ class CfnBucketObject:
 
 
 class CfnBucket:
+    FORCE_OVERWRITE = False
+
     def __init__(self, config: AwsSettings):
         self.bucket_name = config.cfn_bucket
         self.region = config.region
@@ -44,12 +46,16 @@ class CfnBucket:
     # Upload content object to S3 bucket
     def upload(self, object: Uploadable, overwrite: bool = False):
         s3 = self.s3
-        try:
-            s3.head_object(Bucket=self.bucket_name, Key=object.key())
-            # print(f"Key {object.key()} already exists in {self.bucket_name}, not uploading")
-        except botocore.exceptions.ClientError as ex:
-            if ex.response["ResponseMetadata"]["HTTPStatusCode"] == 404:
-                s3.put_object(Bucket=self.bucket_name, Key=object.key(), Body=object.body(), ServerSideEncryption="AES256")
-            else:
-                raise
+
+        if self.FORCE_OVERWRITE:
+            s3.put_object(Bucket=self.bucket_name, Key=object.key(), Body=object.body(), ServerSideEncryption="AES256")
+        else:
+            try:
+                s3.head_object(Bucket=self.bucket_name, Key=object.key())
+                # print(f"Key {object.key()} already exists in {self.bucket_name}, not uploading")
+            except botocore.exceptions.ClientError as ex:
+                if ex.response["ResponseMetadata"]["HTTPStatusCode"] == 404:
+                    s3.put_object(Bucket=self.bucket_name, Key=object.key(), Body=object.body(), ServerSideEncryption="AES256")
+                else:
+                    raise
         return CfnBucketObject(self, object.key())
